Update frontend
1. Every error will be warning attached with response
2. Warning color will be yellow

Tested-by: zlq
